refactor: consolidate 4 duplicate Chrome API mock files into one
- Merge background/popup/options __mocks__/chrome.ts into shared/__mocks__/chrome.ts
- Update test imports to use the consolidated mock
- Fix options test to define mockDocumentElement before global stub

Closes #25
