--- src/background/__mocks__/chrome.ts
@@ -1,74 +1,0 @@
-// Mock Chrome APIs for testing
-import { vi } from 'vitest';
-
-const mockChrome = {
-  alarms: {
-    create: vi.fn(),
-    onAlarm: {
-      addListener: vi.fn(),
-    },
-  },
-  commands: {
-    onCommand: {
-      addListener: vi.fn(),
-    },
-  },
-  runtime: {
-    onInstalled: {
-      addListener: vi.fn(),
-    },
-    onStartup: {
-      addListener: vi.fn(),
-    },
-    onMessage: {
-      addListener: vi.fn(),
-    },
-    getURL: vi.fn((path: string) => `chrome-extension://mock/${path}`),
-  },
-  tabGroups: {
-    query: vi.fn(() => Promise.resolve([])),
-  },
-  tabs: {
-    query: vi.fn(() => Promise.resolve([])),
-    remove: vi.fn(() => Promise.resolve()),
-    onActivated: {
-      addListener: vi.fn(),
-    },
-    onUpdated: {
-      addListener: vi.fn(),
-    },
-    onCreated: {
-      addListener: vi.fn(),
-    },
-    onRemoved: {
-      addListener: vi.fn(),
-    },
-  },
-  action: {
-    setIcon: vi.fn((options, callback) => {
-      if (callback) callback();
-      return Promise.resolve();
-    }),
-  },
-  notifications: {
-    create: vi.fn((options, callback) => {
-      if (callback) callback();
-      return Promise.resolve('mock-notification-id');
-    }),
-  },
-  storage: {
-    local: {
-      get: vi.fn(() => Promise.resolve({})),
-      set: vi.fn(() => Promise.resolve()),
-    },
-    sync: {
-      get: vi.fn(() => Promise.resolve({})),
-      set: vi.fn(() => Promise.resolve()),
-    },
-    onChanged: {
-      addListener: vi.fn(),
-    },
-  },
-};
-
-vi.stubGlobal('chrome', mockChrome);
